构建v1.0 b3

+++ src/com/EasySQL/JDBC_Driver.java
@@ -11,23 +11,33 @@ import com.EasySQL.Exception.StatementException;
 public class JDBC_Driver implements EasySQL{
 	Connection con;
 	Statement stat;
+	boolean login = false;
+	String url_ = "";
 
 	@Override
 	public void Login(String ip, int port, String database, String user,String password) throws ClassNotFoundException, SQLException {
 		Class.forName("com.mysql.cj.jdbc.Driver");
 		String url = "jdbc:mysql://"+ip+":"+port+"/"+database;
+		url_ = "jdbc:mysql://"+ip+":"+port+"/";
 	    con = DriverManager.getConnection(url,user,password);
 		stat = con.createStatement();
+		login = true;
 	}
 
 	@Override
-	public void NormalCommandExec(String command) throws SQLException {
+	public void NormalCommandExec(String command) throws SQLException, StatementException {
+		if(login==false){
+			throw new StatementException("Unlogin!");
+		}
 		stat.execute(command);
 	}
 
 	@Override
 	public ResultSet ResultCommandExec(String command)
 			throws StatementException, SQLException {
+		if(login==false){
+			throw new StatementException("Unlogin!");
+		}
 		stat.execute(command);
 		ResultSet result = stat.getResultSet();
 		if(result == null){
@@ -38,13 +48,19 @@ public class JDBC_Driver implements EasySQL{
 	}
 
 	@Override
-	public void NormalCommandExecQuery(String command) throws SQLException {
+	public void NormalCommandExecQuery(String command) throws SQLException, StatementException {
+		if(login==false){
+			throw new StatementException("Unlogin!");
+		}
 		stat.executeQuery(command);
 	}
 
 	@Override
 	public ResultSet ResultCommandExecQuery(String command)
 			throws StatementException, SQLException {
+		if(login==false){
+			throw new StatementException("Unlogin!");
+		}
 		ResultSet result = stat.executeQuery(command);
 		if(result == null){
 			throw new StatementException("Result is null!");
@@ -56,6 +72,9 @@ public class JDBC_Driver implements EasySQL{
 	@Override
 	public void CreateDatabase(String name, Encoding encoding)
 			throws SQLException, StatementException {
+		if(login==false){
+			throw new StatementException("Unlogin!");
+		}
 		switch(encoding){
 		case ASCII:
 			this.NormalCommandExec("CREATE DATABASE "+name+" DEFAULT CHARACTER SET ascii");
@@ -75,8 +94,32 @@ public class JDBC_Driver implements EasySQL{
 	}
 
 	@Override
-	public void DeleteDatabase(String name) throws SQLException {
+	public void DeleteDatabase(String name) throws SQLException, StatementException {
+		if(login==false){
+			throw new StatementException("Unlogin!");
+		}
 		this.NormalCommandExec("DROP DATABASE "+name);
+		
+	}
+
+	@Override
+	public void CreateTable(String name, String format) throws SQLException, StatementException {
+		if(login==false){
+			throw new StatementException("Unlogin!");
+		}
+		this.NormalCommandExecQuery("CREATE TABLE "+name+"("+format+")");
+	}
+
+	@Override
+	public void CloseAll() throws SQLException, StatementException {
+		if(login==false){
+			throw new StatementException("Unlogin!");
+		}
+		if(!con.isClosed()){
+			con.close();
+			stat.close();
+		}
+		login = false;
 		
 	}
